requirements.txt
altgraph==0.17.5
black==26.1.0
bottle==0.13.4
cabarchive==0.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
filelock==3.20.3
freeze-core==0.5.0
idna==3.11
lief==0.17.1
mypy_extensions==1.1.0
Nuitka==2.8.10
ordered-set==4.1.0
packaging==26.0
pathspec==1.0.3
pefile==2024.8.26
platformdirs==4.5.1
proxy_tools==0.1.0
pycparser==3.0
pyinstaller-hooks-contrib==2026.0
python-dotenv==1.0.0
python-msilib==0.5.0
pythonnet==3.0.5
pytokens==0.4.0
pywebview==6.1
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.10.1
striprtf==0.0.29
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
wheel==0.46.3
zstandard==0.25.0
